src/lambda/markdownsource/requirements.txt
markdown
python-dateutil
pytz

